% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{airtemp}
\alias{airtemp}
\title{Air temperature}
\format{
A data frame with 100 rows and 4 variables:
\describe{
  \item{Unit}{A unit number.}
  \item{DayNo}{Day number.}
  \item{Mercury, Thermistor}{Temperature (°C) measured by a standard glass mercury thermometer and new electronic thermometer.}
}
}
\source{
Welham, S. J., Gezan, S. A., Clark, S. J., and Mead, A. (2015) \emph{Statistical Methods in Biology: Design and analysis of experiments and regression}
}
\usage{
airtemp
}
\description{
Measurements of air temperature (°C) were made at approximately 9 a.m. on 100 days during 2006 with a standard glass mercury dry bulb thermometer and a new electronic dry bulb thermistor probe.
}
\keyword{datasets}
