% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cross}
\alias{cross}
\title{Cross}
\format{
A data frame with 4 variables: \code{Seed}, \code{Female}, \code{Male}, \code{Root}.
}
\source{
Welham, S. J., Gezan, S. A., Clark, S. J., and Mead, A. (2015) \emph{Statistical Methods in Biology: Design and analysis of experiments and regression}
}
\usage{
cross
}
\description{
Description.
}
\keyword{datasets}
