% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demethylation}
\alias{demethylation}
\title{Demethylation}
\format{
A data frame with 4 variables: \code{DTray}, \code{Dose}, \code{Total}, \code{Normal}.
}
\source{
Welham, S. J., Gezan, S. A., Clark, S. J., and Mead, A. (2015) \emph{Statistical Methods in Biology: Design and analysis of experiments and regression}
}
\usage{
demethylation
}
\description{
Description.
}
\keyword{datasets}
