% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elisa}
\alias{elisa}
\title{Elisa}
\format{
A data frame with 3 variables: \code{DUnit}, \code{Concentration}, \code{Absorbance}.
}
\source{
Welham, S. J., Gezan, S. A., Clark, S. J., and Mead, A. (2015) \emph{Statistical Methods in Biology: Design and analysis of experiments and regression}
}
\usage{
elisa
}
\description{
Description.
}
\keyword{datasets}
