% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{latinsquare1}
\alias{latinsquare1}
\title{Latinsquare 1}
\format{
A data frame with 4 variables: \code{Field}, \code{Row}, \code{Column}, \code{Treatment}.
}
\source{
Welham, S. J., Gezan, S. A., Clark, S. J., and Mead, A. (2015) \emph{Statistical Methods in Biology: Design and analysis of experiments and regression}
}
\usage{
latinsquare1
}
\description{
Description.
}
\keyword{datasets}
