% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{latinsquare2}
\alias{latinsquare2}
\title{Latinsquare 2}
\format{
A data frame with 4 variables: \code{Rep}, \code{Stack}, \code{Position}, \code{Treatment}.
}
\source{
Welham, S. J., Gezan, S. A., Clark, S. J., and Mead, A. (2015) \emph{Statistical Methods in Biology: Design and analysis of experiments and regression}
}
\usage{
latinsquare2
}
\description{
Description.
}
\keyword{datasets}
