% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transect}
\alias{transect}
\title{Transect}
\format{
A data frame with 4 variables: \code{DPlant}, \code{Distance}, \code{fDist}, \code{Count}.
}
\source{
Welham, S. J., Gezan, S. A., Clark, S. J., and Mead, A. (2015) \emph{Statistical Methods in Biology: Design and analysis of experiments and regression}
}
\usage{
transect
}
\description{
Description.
}
\keyword{datasets}
