% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{willow}
\alias{willow}
\title{Willow}
\format{
A data frame with 3 variables: \code{DBeetle}, \code{Length}, \code{Width}.
}
\source{
Welham, S. J., Gezan, S. A., Clark, S. J., and Mead, A. (2015) \emph{Statistical Methods in Biology: Design and analysis of experiments and regression}
}
\usage{
willow
}
\description{
Description.
}
\keyword{datasets}
