% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{airtemp}
\alias{airtemp}
\title{Air Temperature}
\format{
A data frame with 4 variables: \code{Unit}, \code{DayNo}, \code{Mercury}, \code{Thermistor}.
}
\source{
Welham, S. J., Gezan, S. A., Clark, S. J., and Mead, A. (2015) \emph{Statistical Methods in Biology: Design and analysis of experiments and regression}
}
\usage{
airtemp
}
\description{
Measurements of air temperature (°C) were made at approximately 9 a.m. on 100 days during 2006 (N = 100) with a standard glass mercury dry bulb thermometer and a new electronic dry bulb thermistor probe.
}
\keyword{datasets}
