% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aphids}
\alias{aphids}
\title{Aphids}
\format{
A data frame with 4 variables: \code{ID}, \code{Field}, \code{Sample}, \code{AphidCount}.
}
\source{
Welham, S. J., Gezan, S. A., Clark, S. J., and Mead, A. (2015) \emph{Statistical Methods in Biology: Design and analysis of experiments and regression}
}
\usage{
aphids
}
\description{
Description.
}
\keyword{datasets}
