% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biomassc}
\alias{biomassc}
\title{Measuring soil microbial biomass}
\format{
A data frame with 5 variables: \code{DSample}, \code{Size}, \code{Weight}, \code{Time}, \code{C}.
}
\source{
Welham, S. J., Gezan, S. A., Clark, S. J., and Mead, A. (2015) \emph{Statistical Methods in Biology: Design and analysis of experiments and regression}
}
\usage{
biomassc
}
\description{
An experiment was done to investigate the effects of changing the procedure for processing samples to obtain measurements of carbon in soil microbial biomass (as mg C per kg soil). The protocol under examination used 200 g soil samples passed over a 2.5 mm sieve and shaken for 60 min. The experiment tested the effects of a larger sieve, two smaller sample weights and a reduced shaking time, giving a 2 × 3 × 2 factorial structure. Each of the 12 treatment combinations was replicated four times in a CRD. The aim of analysis is to quantify the effects of the individual modifications, whether they interact, and to evaluate whether any of the modified procedures obtain results within 10% of the standard protocol
}
\keyword{datasets}
