% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{calcium}
\alias{calcium}
\title{Calcium pot trial}
\format{
A data frame with 3 variables: \code{Pot}, \code{Calcium}, \code{Length}.
}
\source{
Welham, S. J., Gezan, S. A., Clark, S. J., and Mead, A. (2015) \emph{Statistical Methods in Biology: Design and analysis of experiments and regression}
}
\usage{
calcium
}
\description{
An experiment devised to evaluate the effect of four relative concentrations (levels) of calcium (A = 1, B = 5, C = 10, D = 20) on root growth. Each treatment was applied to five individual plants growing in pots. The experiment used a CRD, as shown in Table 3.2, and measurements of total root length (cm) were made on pots 1–20 (in order) at the end of the experiment.
}
\keyword{datasets}
