% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{competition}
\alias{competition}
\title{Competition}
\format{
A data frame with 7 variables: \code{ID}, \code{Block}, \code{WholePlot}, \code{Subplot}, \code{Irrigation}, \code{Species}, \code{Grain}.
}
\source{
Welham, S. J., Gezan, S. A., Clark, S. J., and Mead, A. (2015) \emph{Statistical Methods in Biology: Design and analysis of experiments and regression}
}
\usage{
competition
}
\description{
Description.
}
\examples{
library(desplot)
desplot(Irrigation ~ WholePlot + Subplot | Block, data=competition, col=Species, cex=2)
summary(aov(Grain ~ Species*Irrigation + Error(Block/WholePlot/Subplot), data=competition))
}
\keyword{datasets}
