% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cotton}
\alias{cotton}
\title{Cotton}
\format{
A data frame with 4 variables: \code{ID}, \code{H}, \code{I}, \code{Weight}.
}
\source{
Welham, S. J., Gezan, S. A., Clark, S. J., and Mead, A. (2015) \emph{Statistical Methods in Biology: Design and analysis of experiments and regression}
}
\usage{
cotton
}
\description{
Description.
}
\keyword{datasets}
