% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{forest}
\alias{forest}
\title{Forest}
\format{
A data frame with 4 variables: \code{DPlot}, \code{Type}, \code{SD}, \code{QD}.
}
\source{
Welham, S. J., Gezan, S. A., Clark, S. J., and Mead, A. (2015) \emph{Statistical Methods in Biology: Design and analysis of experiments and regression}
}
\usage{
forest
}
\description{
Description.
}
\keyword{datasets}
